Extract UserSelection from __init__
